+++ src/flow-view.js
@@ -1,6 +1,3 @@
 import Canvas from './Canvas'
-import actions from './actions'
-import components from './components'
-import reducers from './reducers'
 
-export { Canvas, actions, components, reducers }
+export { Canvas }